// Richbourn-Dashboard/src/app/core/models/item.model.ts
export interface ItemModel {
    SlNo: any;
    id: number
    Item: any;
    ItemCode: any
    ItemName: string;
    BaseUnit: string;
    Category: string;
    SubCategory: string;
    ItemDesc: string;
    Cost: string;
    MRP: string;
    ReOrder: string;
    HsnCode: string;
    BaseQty: string;
}